refactor: use simple middleware
