--- types/middlewares.ts
@@ -1,24 +1,0 @@
-import type { NextRequest, NextResponse } from 'next/server';
-
-export type NextMiddlewareLocale = {
-  isLocalisedRoute: boolean;
-  currentLocale: string | null;
-  availableLocales: string[];
-};
-
-export type CustomMiddleware = {
-  handler: (
-    request: NextRequest,
-    response: typeof NextResponse,
-    locale: NextMiddlewareLocale
-  ) => Promise<NextResponse>;
-  matcher: (request: NextRequest, locale: NextMiddlewareLocale) => boolean;
-  routes: (string | RegExp)[];
-};
-
-export type RouteSegment = {
-  filename: string;
-  localised: boolean;
-  routeWithLocale: string;
-  pathname: string;
-};
